leaf80: modularize macro_expand_v0 internals
